SAK-31603 Correctly setup OAuth for upgrades.
On upgrade installations we weren’t creating the OAUTH_* tables, this meant that the oauth functionality would fail.
This uses the old hibernate3:hbm2ddl maven plugin to get the same DDL statements and needs editing to run on different schemas but this is just here as a reference as it shouldn't be needed again.