Provide Toggle Filter to Only Display Updated/New Items (#3157)
* #1396 if a comment column is upload but there is no corresponding gb item, return an error

* Fix the condition
Remove errors if valid file uploaded

* #3059 allow points to be changed in an import file
Adds a new method to the GBNG business service to get assignment by name. Adds a new status of modified. Shows the list of items to be modified on the confirmation page and then goes through and updates them.
Added some javadoc to the other statuses.

* #3131 Provide a toggle to hide items with no changes

 Conflicts:
	gradebookng/tool/src/java/org/sakaiproject/gradebookng/GradebookNgApplication.properties