Fix error when importing grade items with extra credit (#3158)
* #1396 if a comment column is upload but there is no corresponding gb item, return an error

* Fix the condition
Remove errors if valid file uploaded
#1396

* #3059 allow points to be changed in an import file
Adds a new method to the GBNG business service to get assignment by name. Adds a new status of modified. Shows the list of items to be modified on the confirmation page and then goes through and updates them.
Added some javadoc to the other statuses.

* #3131 Provide a toggle to hide items with no changes

* #3129 allow items that are extra credit to be imported
Also replaced some strings with proper enums and removed some of the unnecessary PropertyModel code when a value can just be wrapped in a Model instead, where needed.
Cleaned up a few other Enum types to make it clearer what they are, i.e. a comment column is a standalone column, not attached to a grade item, so not require to call it grade_item_with_comments.