SAK-31656 Override defaults first in SCSS skin.
Locally at Oxford we've got our own changes to the Sakai default skin and these are in a customisations file. By default this customisation file in pulled in at the end of the _defaults.css file so you can override any of the variables in _defaults.css. As an example in _defaults.scss there is a primary colour defined as:
$primary-color: #2a94c0 !default;
we then in our customizations file re-define the primary colour to be:
$primary-color: #002147;
the problem is that if this primary colour is referenced anywhere else in _defaults.scss our re-define doesn't work because the value is already copied into the new variables, as an example:
$tool-menu-current-color: $primary-color !default;
here the tool menu current color will be the standard Sakai one, rather than our Oxford one. The !default on the end of the line says to only assign a value if one isn't already present so it seems that the sensible option is to import the customisations at the start of the defaults file rather than the end so that re-definitions take effect in all locations in the file.