SAK-31682 autocomplete suggestions for hosts (PA Tool). (#3197)
Output the known list of hosts so that they can autocompleted by browsers that support them, This doesn’t stop people from typing in hosts that don’t exist in the dropdown but helps prevent mistypes of existing hosts. The in browser autocomplete doesn't work after the first hostname has been entered but it's a simple solution for the most common case of a message for a specific host.

The <datalist> element is supported in most popular browser on the desktop apart from Safari. When it isn't supported it's just ignored so it doesn't break anything.

http://caniuse.com/#feat=datalist