SAK-31740 Don’t specify http methods, apply to all.
If you specify a limited number of http-methods then tomcat warns on startup that unlisted methods will be allowed through without any security constrains applied to them.

All the methods in DavServlet were in the list so applying the security constraint to all methods won’t change the behaviour.

The error that would appear in the logs was: 

    For security constraints with URL pattern [/*] only the HTTP methods [HEAD MKCOL POST PROPFIND LOCK COPY OPTIONS PUT DELETE MOVE GET PROPPATCH UNLOCK] are covered. All other methods are uncovered.