WL-4471 Switch lessons to strict caching.
When you have non strict caching even if there is only one thread updating the row if you have multiple other threads reading that row at the same time it appears to result in old data coming back from the cache. When Lessons adds an item to the page it does one INSERT and 2 UPDATES and it does these in different transactions. We would see the state appear in the cache of the first INSERT but not the change from the UPDATES.

The second thread doing the reads was the search indexer which would load the recently added item.