#3274 Refactor to support the enhanced import item display
Get rid of ProcessedGradeItemStatus finally, and use an Enum as it should always have been.
Simplify ProcessedGradeItem to map any updated values back into the itemTitle rather than the assignmentTitle field and have to check both.
Improve the filtering on the selection page to use streams.
Added in a mapping so that the item and corresponding comment column can be pulled out at once
Fixed a bunch of defaults when setting up objects as it just made sense.
Have removed the display of the comments for now
Updated tests