SAK-31773 Disable last bits of PA on mobile.
The PA tool doesn’t currently render correctly on a mobile and until this is fix it should all be hidden on a mobile so that it doesn’t interfere. with the normal use of the interface. In the future we want to fix the PA announcements so they work well on a mobile.