SAM-2970 Refactor Extended Time
Front end is working. Jquery date selector only for new items.

When saving the extended time entries, all new entries are compared
to what is in the database by ID. If there are
no matches than (that is an old entry is not present in the new ones),
than that entry is removed. (see saveExtendedTimeEntriesHelper within
ExtendedTimeFacade.java).

ExtendedTimeDeliveryService is used as a medium for the delivery
aspect of the assessments. This has been modified to use the new
ExtendedTime entries. This should reduce the testing load necessary.

Entries are validating when an assessment is published / all ready
published exam is saved.

All done except for auto-submit. Minor testing done

KNOWN BUG (probably all ready present):
When viewing the submissions on the totalScores.jsp page, assessments
will be marked LATE even though they where submitted within the users
extended time window.