LSNBLDR-709 Added calendar component for Lessons page.
New option 'Embed Calendar' is added in the list on Lessons page.
SimplePageItem of type 'Calendar' created for calendar widget.
Calendar shows events from all the merged calendars in the Lessons tool.

Exposed method 'getCalendarReferences' in CalendarService to get
references from all calendars for a site.
Added attributes 'description' and 'eventImage' in the EventSummary class
which will be used in the json returned by CalendarEventEntityProvider.
Created map for event type and image in the CalendarUtil class which can be
accessed using method 'getEventImageMap'.