SAK-31705 Added 'Search Users' facility to ste-info. (#3240)
Added Search input and clear search in the chef_site-siteInfo-list.vm.
If search word is in session state then Participant List is filtered for
the search term in each Participant's displayName and displayId.
Corrected the position of search box in the participant's table.

Have added new variable to STATE_SITE_PARTICIPANT_LIST in SiteAction class
to save the original site members list. Search is case insensitive.