SAK-31404, SAK-31405 and SAK-31406 : Don’t require a candidate details provider.
Now we make the provider optional. I did look at making the field Optional<CandidateDetailsProvider> but it needs more work. This is desirable so that it’s clearer to all callers of methods on this field that they need to handle the case when it’s not set (null).

It would be better to hide all this behind the API so that the tool doesn’t deal with it at all.