LSNBLDR-633 Restricted editing of Lessons subpages to one or more people.
New column 'isOwnedPage' is added in the table 'lesson_builder_pages'.
Similarly new attribute with name 'isOwnedPage' is added in SimplePageImpl
class. This will be set to true for the pages which are assigned to
specific owner.

Change owner dropdown is added in the permissions screen of a Lessons page
so that it can be assigned to maintainer/contributer in the site.
New method 'isStudentPage' is added in 'SimplePageBean' class to check if
a page is student page or not.This method is used in another producer
classes to check for a student page.