SAK-31405 Allow forcing of anonymous assignments.
Now we make the provider optional. I did look at making the field Optional<CandidateDetailsProvider> but it needs more work. This is desirable so that it’s clearer to all callers of methods on this field that they need to handle the case when it’s not set (null).

It would be better to hide all this behind the API so that the tool doesn’t deal with it at all.

Don’t log at warn but at debug when users aren’t found as this is an expected case.

Also switched ValueEncryptionUtilities to be a service so that it’s easier to switch to a static key if we need to in the future for performance reasons.
Add padding to values so that all encrypted values will be the same length.