LSNBLDR-727 Made LB-CSS folder hidden by default.
In SimplePageBean class when LB-CSS folder is added in resources, the
PROP_HIDDEN_WITH_ACCESSIBLE_CONTENT property is set to true for the folder.

For the existing LB-CSS folders in database a new job(class) -
LBCSSFolderHideJob is added. When this job is run , it will hide all the
visible LB-CSS folders from access users.