SAK-31836 Remove unused template file in site-info
While looking at how LTI tools were displayed I found that there are no references to this file in the codebase and I can’t even find any references to it in the history. It came in as party of SAK-16600 but doesn’t look to have ever been used.