SAK-31836 Don't preselect LTI tools in site (#3455)
* SAK-31836 Remove unused template file in site-info

While looking at how LTI tools were displayed I found that there are no references to this file in the codebase and I can’t even find any references to it in the history. It came in as party of SAK-16600 but doesn’t look to have ever been used.

* SAK-31836 Don’t pre-select LTI tools in Worksite Setup

When editing an existing site through worksite setup if the tool is already placed in a second site then it gets automatically added when editing any other site.