SAK-31819 Move ScheduledInvocationManager to directly use Quartz (#3420)
* Remove scheduled invocation runner.

* Update with a working hibernate test.

* Update to use both opaqueId and componentId.

This means we don’t have to rely on opaqueId being globally unique. Doesn’t yet full work.

* Use ScheduledInvocationManager.deleteDelayedInvocation

In 4 places code was getting a list of all delayed invocations and then for each one removing it, this functionality is available directly in the API and so the code can be simplified.

* Refactor tests to get working again.

* Work on migration.

* Time -> Instant on ScheduledInvocationManager.

The Time class is deprecated so provide an alternative using the newer JDK 8 Instant class.

* Fixup the migration job so it works better.

* Fix spring wiring.

* Register our listener on the scheduler.

* Fix startup in Sakai

Wire up all our beans correctly in Sakai. Correctly remove context_mapping entries when we remove something before it’s fired. Use the standard Sakai table naming of an underscore between parts. Added test of the new remove method.

In Sakai there are 3 Sakai session managers (help’s search, sitestats’s external DB and the main one) so we need to be explicit about which one we want.

* WIP needs checking

* Remove classes that aren’t used.

* Cleanup BaseMessage (no functional changes).

* Added a validation job that checks the state.

This jobs checks to see that the DAO and the quartz data are in sync. This is useful for debugging any issues that may happen. Also removed out some commented out code.

* SAK-31677 Make the service transactional.

Make the published API of the ScheduledInvocationManager transactional. The trigger will be transactional using it’s DAO.