SAK-31905 Use the GroupProvider when packing/unpacking.
Use the provider for packing/unpacking of the provider IDs. The main change for this fix is in SiteManageGroupSectionRoleHandler where we use the GroupProvider rather than the SiteGroupHelper for packing the IDs.

Added generics to RoleGroupEventWatcher as I was looking at what the code was doing, not other changes in that file.

I inlined the `getProviderString(List<String>)` call as it’s only used once and wanted to make it clearer which call should be used in which place.

The packing (and unpacking) that should be used is

`GroupProvider.packId()` for `Group.setProviderGroupId()`
`SiteGroupHelper.pack()` fot `SiteContants.GROUP_PROP_ROLE_PROVIDERID`