SAK-31957 Don’t fail on sorting with deleted users.
If a site has an announcement made by a user who is now deleted then attempting to sort by the author results in an error as the anonymous
user doesn’t have a sort name.

Refactor comparator into it’s own class.
This also refactors out the group comparator into it's own one, so the
comparators can be type safe.
Also removed sortbygrouptitle and sortbygroupdescription as there is
not use of these options and they were resulting in the comparators not using stricter types.