SAK-31957 Don't NPE when announcements sorted by user. (#3613)
* SAK-31957 Add Generics to MessageChannel interface

This makes it easier to add generics to classes that implement these interfaces and so check that our types are correct.

* SAK-31957 Added generics the to AnnouncementChannel

This makes it easier to see what types the calls should be returning.

* SAK-31957 Don’t fail on sorting with deleted users.

If a site has an announcement made by a user who is now deleted then attempting to sort by the author results in an error as the anonymous
user doesn’t have a sort name.

Refactor comparator into it’s own class.
This also refactors out the group comparator into it's own one, so the
comparators can be type safe.
Also removed sortbygrouptitle and sortbygroupdescription as there is
not use of these options and they were resulting in the comparators not using stricter types.

(cherry picked from commit e06c80b155721354d5926e642a2f42ca2ff47483)