3631 Improve link passthrough in GradebookNG
Within the GradebookNG tool, there is a fixed column which appears only
once scrolled to the right. Clicks on elements within that column (i.e.
student name / the row itself) would not be passed through to the actual
element where the event is processed. This fixes that so that the
clicks are passed through correctly, for example: allowing one to
access the student overview panel when scrolled to the right.

This change also removes the scroll left aspect of the clickthrough.