SAK-32263 Preserve Forum Topic Order on Dupe
When a forum was being duplicated via transferCopyEntitiesRefMigrator
the topic List was retrieved in an unsorted order. As such, it was
possible for a topic with a sortOrder > 1 to be selected first and
imported into the new forum. Then when the forum was saved in
    MessageForumsForumManagerImpl#saveDiscussionForum
      (org.sakaiproject.api.app.messageforums.DiscussionForum, boolean,
        boolean, java.lang.String)
the topics would be retrieved using BaseForumImpl#getTopics. This
method would first check to see if the topicSet was sorted by going
through the topics and checking their sort order, starting at 1.

With the topic list being retrieved in an unsorted order, you could
have the first topic inserted to have a sort order > 1. Then when
getTopics was ran, with only one topic in the set (whose sortOrder
is > 1), it would "correct" the sortOrder of this topic.

This change causes the topics to be retireved in a sorted order, thus
the first topic inserted would have a sortOrder of 1 and getTopics would
not rewrite the sortOrder.