3651 In GradebookNG prevent erroneous row added
Within the GradebookNG tool if one where to turn "Group by Category" on
(when categories where enabled) and then to turn off Categories, if the
user where to scroll right then the fixed header would have an
extraneous row on it. This change causes the row to only be added when
group by category is enabled.