SAK-31944 Fix Calendar tool Month display name (#3582)
* SAK-31944 Fix Calendar tool Month display name

Previously in the Calendar tool, the month name would be displayed
incorrectly when it was the 31 day of a month. This was resolved
by using the new Java 8 YearMonth object.

Minor cleanup in CalendarUtil.java

* SAK-31944 Add test for month calculations on 31st.

When it’s the 31st of the month the calculated day of the month is shown incorrectly.