SAK-32151 Sakai Rubrics initial support. This includes a Rubrics Service API and Web Components, as well as a Sakai Rubrics Tool for managing Rubrics.
This commit includes the basis for a very minor 3rd party Polymer library
modification. The file is contained in:
rubrics/services/src/main/webapp/lib/polymer-sortablejs/polymer-sortablejs.html
This is the community version that is pulled down as part of the build.
A subsequent commit will apply changes to fix an issue with that
library.