SAK-31405 Allow institution provided anonymous IDs (+ additional candidate notes) (#3331)
* SAK-31404, SAK-31405 and SAK-31406

anonymous coursework project

- Manage Candidate ID and additional notes.
- Sample, html, zip and more

* Add support to LDAP multi-valued properties

* SAK-31404, SAK-31405 and SAK-31406 : Move providers out of the kernel.

The kernel shouldn't depend on anything else in Sakai.

* SAK-31404, SAK-31405 and SAK-31406 : Don’t require a candidate details provider.

Now we make the provider optional. I did look at making the field Optional<CandidateDetailsProvider> but it needs more work. This is desirable so that it’s clearer to all callers of methods on this field that they need to handle the case when it’s not set (null).

It would be better to hide all this behind the API so that the tool doesn’t deal with it at all.

* SAK-31404, SAK-31405 and SAK-31406 : We expect some users to not be found.

Don’t log at warn but at debug when users aren’t found as this is an expected case.

* SAK-31404, SAK-31405 and SAK-31406 : Add the logger back in so it compiles.

* SAK-31404, SAK-31405 and SAK-31406

- Switched ValueEncryptionUtilities to be a service so that it’s easier
to switch to a static key if we need to in the future for performance
reasons.
- Add padding to values so that all encrypted values will be the same
length.
- Added MappingDetailProvider

* SAK-31405

* additional notes added to :
- submission grading screen (only for no group assignments)
- grade report screen : html table
- grade report screen : xls export
- submission list -> download all -> grades file inside zip (csv, xls)

* candidateDetailProvider is no longer injected in vm files (also site).
Now we use notesMap

* SAK-31405 Allow forcing of anonymous assignments.

Now we make the provider optional. I did look at making the field Optional<CandidateDetailsProvider> but it needs more work. This is desirable so that it’s clearer to all callers of methods on this field that they need to handle the case when it’s not set (null).

It would be better to hide all this behind the API so that the tool doesn’t deal with it at all.

Don’t log at warn but at debug when users aren’t found as this is an expected case.

Also switched ValueEncryptionUtilities to be a service so that it’s easier to switch to a static key if we need to in the future for performance reasons.
Add padding to values so that all encrypted values will be the same length.

* SAK-31405 Make the sample data display nicer.

This allows wrapping in the interface and doesn’t make it unusable.
Also stop the NPE by trying to encrypt null.

* SAK-31405 Fix indentation on API

* SAK-31405 Added more documentation and a readme.md

This is to make it easier to understand how this provider is designed to work.

* SAK-31405 Switch the SLF4J from commons-logging.