LSNBLDR-790 Checklist Improvements
* Improve view Other Checklists progress sorting

* Trim whitespace leading and trailing whitespace from checklist titles
and checklist item names. Also error when attempting to save an itiem
with an empty title (or all spaces).

* Checklist items may now link to other items on the same page and
automagically check themself if the item they have linked to has
been completed. If the item linked to is not visible, the checklist
item is blurred on the page via CSS magic (and the title is changed
to "Not available".

All required items may be linked to. Within a checklist, one may only
link to a unique item on the page. Users may not check items which
are linked to another item.

Add a popup and visual indicator that a checklist linked item is linked
to another item.

Linked checklist styling has been improved and the tooltip messaging
now reflects the status of the checkbox (that is a complete linked item
has a separate message from an incomplete item).

When editing a checklist, the linked items are checked to ensure that
the linked item is on the page.

Only able to link to items which are required. If there are no items
required on the page, than linking is hidden.

* Removed unused imports in ShowPageProducer.java

* Javascript removed from the .html page (the rsf template).