SAK-32185 Circular dependencies check later in build (#3913)
If you attempt todo `mvn validate` againast an empty repository it will fail because inter project dependencies aren't yet installed in the local repository and so can't be found. Moving the circular dependencies check to later in the builds means we can keep doing it but not break the validate phase.