SAK-32127 If making the Resources tool invisible, hide root (#3811)
* SAKAI-32127 If making the Resources tool invisible, hide root

Hiding the page the resources tool is on now sets the root collection
to hidden but users still able to access.

This behavior is allowed by setting the sakai.property
    poh.resources.hidden.special.hidden
to true. The default for this property is false.

* SAK-32127 elfinder now drops users into root of site

elfinder now drops users into the root of a site, instead of into
the resources folder of the site. Changes to builder.js where made
to be compatible with npm 4.0.5.

* SAK-32127 Add Atachments respects Special Hidden

Add Attachments, from Assignments / Forums, now respects when the
siteCollection (aka the root resources folder of a site) is special
hidden (hidden, but contents accessible). This is done by logging the
PermissionException which occurs and continuing.

* SAK-32127 Minor improvements

Minor improvements as a result of code review.