SAK-25750 Upgrade to a newer version of castor.
This means we’re keeping castor in sync with the version pluto depends on. The dependency isn’t needed in the service as it’s pulled in transitively.

This also fixes the startup warning that were being seen with newer versions of Tomcat as the newer castor JAR doesn’t have a META-INF/MANIFEST.MF that contains a Class-Path attribute.