SAK-32150 Allow non-standard site types to work. (#3881)
When using the edit tools menu in Site Info of a site that doesn’t have a standard type set (e.g. not project or course) you are presented with a list of possible tools to add to your site, but when you select some and proceed through the wizard it doesn’t actually add them to the site.

This is because when showing the list of tools available it falls back to the default site (project), but when actually adding them it doesn’t do this.