SAK-32252 Update all mockito verions in pom.xml
This defines a standard mockito version in master/pom.xml and has
everything use that. There are a few tools that use powermock which
isn't yet updated to use the newer mockito 2.x versions in an official
release so until a newer version of powermock comes out I've just
dropped those projects back to the older mockito version.