SAM-3167 Fix for samigo attempting to work early.
The importing job was attempting to read archives containing samigo assessments early in the startup and this showed that the fix to block samigo code until the webapp has been started up wasn’t working correctly.

This correctly uses the `waitLock.wait()` inside the synchronized block, before we were getting an java.lang.IllegalMonitorStateException when attempting to `wait()`;

Also simplified the code as we don’t need to have 2 contexts as we can just actually have one and use more generic interfaces.