SAK-32292 Only run yuicompressor once. (#4038)
Rather than running the yuicompressor twice (bound to difference phases. We just run it once and have all the generated files inside the /target folder instead of inside the source directory.

We also switch the order so that all our error messages are for the original source files rather than for the concatenated ones which makes fixing them much easier.