SAK-32303 Allow long running jobs to be stopped.
Some of the long running jobs can cause problems so allow them to be stopped early. There’s no guarantee that stopping will work as it relies on the job still progressing and being able to check the flag. But it’s helpful where deployments have lots of content and running these jobs can take a long time.