SAK-32316 Allow email/AID for group CSV imports.
When importing a group into a site using a CSV file as well as searching for a user by EID it can be helpful to also search by AID and email address.

We are also more flexible and ignore rows that have empty values in them.
Also delete en_GB properties todo with importing that are identical to en.