SAK-32264 Fix for MIME detection with path has dots (#4061)
* SAK-32264 Detect file extension when site ID has .

If the site ID contains a period ('.') then Sakai incorrectly detects
the file. This fixes that problems.

We don’t need the null check as id should never be null and if it is the Validator call will cause it to blow up anyway, so may as well blow up here.

Split the tests into smaller ones so if a failure occurs it’s more precise and the other tests also run. Also we don’t need all of Sakai kernel up to run this test so switch away from SakaiKernelTestBase.

Dropped the reference to the SAK in the code to keep it clean.