SAK-32349 Use a new transaction for changes.
This is a good solution because the calling code (in assignments) had it’s own hibernate transaction already established. The quartz code commits it’s transaction when it returns and so then second call through to quartz wouldn’t find the trigger that had just been deleted.

So the fix to have the ScheduledInvocationManager use a new transaction for each call so it remains in sync with the quartz datastore.