KNL-1508 Don’t lose title changes on save.
When saving the placement the SQL code calls getTitle() which checks if the placement has a custom title and unless it does just returns the default title for the tool.

This change means that when we changes the title the flag indicating that it has a custom title is also updated and so the save works.