SAK-32360 Always pass site ID to LTI calls.
The LTIService now always wants the site ID passed into all it’s calls as assuming that the context is correct doesn’t work when the tool is in helper mode in My Workspace or the Administration Workspace.

As lots of the method signatures were changing I also did a cleanup, removing lots of useless JavaDoc and adding @Override to the appropriate methods.