SAL-32391 Better warning when ID length exceeded.
When uploading a file of creating a new file that will exceed the length of the ID column the error message generated wasn’t very good (just generic error, of null in message when creating text file).

Now we say what went wrong and when uploading a file specify what the actual limit is so the user has more of an idea how to fix it.