SAK-32151 Sakai Rubrics initial support (#3883)
* SAK-32151 Sakai Rubrics initial support. This includes a Rubrics Service API and
Web Components, as well as a Sakai Rubrics Tool for managing Rubrics.

This commit includes the basis for a very minor 3rd party Polymer library
modification. The file is contained in:
rubrics/services/src/main/webapp/lib/polymer-sortablejs/polymer-sortablejs.html
This is the community version that is pulled down as part of the build.
A subsequent commit will apply changes to fix an issue with that
library.

* SAK-32151 Modification to the polymer-sortablejs.html to fix a bug in the library
which affected proper drag and drop sorting behavior.

This modified file is copied over the 3rd party community version during
the build process. The previous commit contained the unmodified
community base version for this library, so this commit shows the
trivial extent of the customization.

* SAK-32151 Sakai Rubrics initial support. This includes a Rubrics Service API and
Web Components, as well as a Sakai Rubrics Tool for managing Rubrics.

This commit includes the basis for a very minor 3rd party Polymer library
modification. The file is contained in:
rubrics/services/src/main/webapp/lib/polymer-sortablejs/polymer-sortablejs.html
This is the community version that is pulled down as part of the build.
A subsequent commit will apply changes to fix an issue with that
library.

* New rebase in rubrics. Pom changes