SAK-32445 Fix Assignments Grading with ".X" Grades (#4250)
When an assignment was saved with a grade such as ".7" with less
than the number of allowed decimal places (2 is default), the grade was
saved as "0.07"). This changes causes it to be saved as "0.70" by
padding entered scores with trailing 0s as appropriate.
(cherry picked from commit 4ffb8957653d7c3cbf98ad6666742a1982737ace)