SAK-32508 Cache failed lookups for a shorter time.
Also added support for recording an error about why the failure to load an entry occurred, although this isn’t yet shown in the UI.
Removed the ExternalCalendarSubscriptionService cover and updated to use the service, there wasn’t any use of the cover outside the calendar tool.
Split out the preloading of calendars into it’s own bean as this makes testing of the calendar service easier.
Add generics and update to Java 8 for parts of the code I was looking at.
Also drop needless methods from API.