SAK-32440 Don’t fail when creating site with LTI tool
If you attempt to create a new site and include an LTI tool on that site which requires some options to be selected then it would blow up with a NPE exception. This fixes it so that it checks and uses a random UUID when there isn’t yet a current site ID. A random UUID works fine as it means that site tools aren’t shown but system wide ones are. As the site doesn’t exist yet there couldn’t be any site ones yet.