SAK-32427 Fix for user email/AID lookups.
It also fixes a problem with non-id members not being correctly added to the group. If you attempt to add a user with an email address or AID then they get looked up in the earlier part of the wizard, but then not added to the group in the last step.

Also the saveGroupMembership call is redundant and can be removed as saving the site will save any changes to the group membership.