SAK-32552 Allow site members to use /direct/lti/
This allows site members to use /direct/lti, previously you needed to be a maintainer to access these endpoints which mean it’s wasn’t possible to find the LTI tools installed in a site using a JSON request. It just changes the /lti/tool/ /lti/tools /lti/content/ /lit/contents/ endpoints.

It also makes sure we never output passwords in the JSON, even for admins.