SAK-32561 Forward creditor smiley images.
We send all requests to the current webjar path.

We have an additional rule that at the moment is redundant but when ckeditor is next upgraded will mean that the old links continue to work. Once ckeditor has been upgraded we can add FORWARD back to the filter dispatchers that are used, but at the moment while the versions are the same it results in a endless loop.