SAK-32149 elfinder allow upload when you can’t modify
If you have permission to create new files in a folder but you can’t modify the actual folder then we shouldn’t say that you don’t have write access as elfinder then doesn’t allow you to upload a file.

However if you can’t change a folder and you attempt to rename it you will then get an error, but with the permission model that elfinder currently has this is the best we can do.