SAK-32588 Allow uploads without dots in name (#4468)
* SAK-32588 Allow uploads without dots in name.

This allows the upload to work sensibly when there isn’t a dot in the filename.

* Stop logging warnings when things don’t exist.

If the resources tool has been added to the site but the root collection for the site hasn’t yet been created  elfinder will attempt to discover information about a collection that doesn’t exist. This would result in lots of stack traces that aren’t serious so we now ignore them.