SAK-31515 Don’t display Resources when tool missing (#4471)
In elfinder even if the Resources tool was missing from the site the Resources tool would be opened when using elfinder as a file picker. This stops that and prevents the Resources folder form being displayed even when it’s set as the initial folder.

It was always the case that when browsing an alternative site that if the Resources tool wasn’t present you wouldn’t see that folder present in elfinder.

If the resources tool is added, some files added and then removed you will still see the Resources folder when browsing the current site.