SAK-32613 Fix for the last tool in list missing border. (#4507)
The last tool in the list would be missing it’s border. You would only see this problem when you had that tool added to the site and went back to the manage tools page. Typically this is the wiki tool, unless you also have some LTI tools available in the site.