SAK-32625 Don’t fail when site has deleted users.
Calls to /direct/roster/site/….json would fail if the site contained a deleted users. This fixes it so that they aren’t included in the results, just like inactive users.

I also downgraded the logging from info to debug for failure to find users as in production this is an expected case.

I also changed handling of a group not being found, rather than returning an empty collection, it throws an error which gets transformed to a HTTP 404 error. This matches the error handling when a site isn’t found.