SAK-32623 - The date settings of the chat channel cannot be unset (#4519)
The date settings of the chat channel cannot be unset.

Server side: It always require a valid value in the hidden field, but
empty dates are allowed.
Client side: When someone clears the input field, the hidden value is
never updated, and it's processed in the server side. This only affects
the pickers with allowEmptyDate option set to true and the ones who
process the hidden value in the server side.