SAK-32647 Allow CORS checks for all font files.
Lessons serves up HTML pages from resources with a stricter content security policy that the standard /access view. This causes the browser to check for a CORS header on the font files and refuse to load them when it can’t find the header. This allows request from anywhere for the standard font file extensions. We could be more restrictive and only allow it from known hostnames, but this is more error prone and these fonts are available on CDNs anyway.