SAK-32181 - Drop box / email sent when copying files (#4556)
Dropbox is sending emails when files are copied. there are two
preferences, tool and user. The tool's preference is totally ignored and
only the user's is considered.

The move operation doesn't send any notification, copy should do the
same.

This should be discussed with the core team, i think it's not the best
solution.