SAK-32541 Fix handling of expired cache entries. (#4545)
* SAK-32541 Fix handling of expired cache entries.

If you perform a search in the profile tool and then wait for the cache entry to have expired and attempt to view the search page again, or perform another search the profile tool falls over as it expects to get an item from the cache and when it’s returned as null due to being expired it NPEs. There are several stack traces that you can end up having but they appear to resolve around the search cache.

This changes means we don’t bother testing to see if an item is in the cache and just try to get it and check if it’s null. While looking at this I also added generics to the caching to reduce the amount of casting.

* SAK-32541 Update to SLF4J logging call.