SAK-32718 Correctly shutdown the portal chat EB (#4627)
When the portal chat code was getting shut down it’s destroy method wasn’t getting called so it wasn’t unregistering itself from the cluster and this combined with the problem of node failure detection not working was causing new nodes not being able to connect to the cluster. Although it shouldn’t be needed it’s cleaner to un-register from the cluster so we’re putting this back in.