SAK-32758 Add the sakai-ws endpoint and tool
This is a new user picker for adding users to a site. It consists of a
JS frontend and JAX-RS backend. The build is still done using maven but
it uses node to bundle up the frontend JS. The presentation code is all
built around VueJS to allow a more componentised approch to development.