SAK-32755 Allow config of tool to add users (#4679)
This allows the tool used to add users to a site to be configured as a
deployment and a per site level through sakai.properties and site
properties. The site property overrides the sakai.properties config. The
property key is:

    sitemanage.add.user.tool

When looked up it's also checked that it is a valid helper tool to avoid
situations where the add user helper can't be run.