SAK-33419 Don’t die when re-ordering announcements
The issues is that although the session state always had AnnouncementWrappers in it and the comparator works fine on an AnnouncementWrapper when re-ordering the comparator is used on a plain AnnouncementMessage. This would cause a ClassCastException.

The problem with the original refactor was that although I applied generics to some of the code I didn’t go far enough which is why I missed the reordering case. A simple fix would be to just drop the generics which would fix the problem but I think in the long run we’re better to try to keep the generics as it makes the code more maintainable and easier to refactor later on.

This adds generics to all the uses of AnnouncementComparator so that we’re sure there aren’t any more cases of the comparator being used on the wrong type. We also rename the Comparator to make it clearer what it will sort.