SAK-33551 Refactor PDF code from CalendarService
The PDF calendar creation code is very self contained and can be held in
an external class and called on by the main BaseCalendarService. This
makes the code easier to test as well as better following the single
responsibilty principal.

Tests are also added that check that PDF generation is working and while
we're not doing good checks of what we render. It does make development
easier as you can test the PDF code without re-deploying Sakai.