SAK-33498 Removed unused commons-email dependency (#4904)
This dependency is no longer used and just increases the size of the
build. This probably remains from when mailsender used to send out email
directly instead of using the kernel's email service.

The activation dependency can also be removed as it isn't directly used
by the code and is correctly pulled in as a trasative dependency.