SAK-33551 Upgrade to newer version of Apache FOP.
With the newer version of Apache FOP changes were needed to the XSL
files to produce the XSL-FO. The file was badly indented and so got
re-indented here to make it easier to read and understand what was going
on.

Font loading also changed so the metrics and loading mechanism were
changed. We also only create the FOP Factory once now, rather than on every PDF creation.