SAK-33551 Upgrade Apache FOP in Calendar. (#4955)
* SAK-33551 Refactor PDF code from CalendarService

The PDF calendar creation code is very self contained and can be held in
an external class and called on by the main BaseCalendarService. This
makes the code easier to test as well as better following the single
responsibilty principal.

Tests are also added that check that PDF generation is working and while
we're not doing good checks of what we render. It does make development
easier as you can test the PDF code without re-deploying Sakai.

* SAK-33551 Upgrade to newer version of Apache FOP.

With the newer version of Apache FOP changes were needed to the XSL
files to produce the XSL-FO. The file was badly indented and so got
re-indented here to make it easier to read and understand what was going
on.

Font loading also changed so the metrics and loading mechanism were
changed. We also only create the FOP Factory once now, rather than on every PDF creation.