SAK-33564 Don’t load all sites in the system. (#4965)
In order to populate the department dropdown the LTI Admin tool was loading all sites in the system. The fix is to no load this dropdown and just have a text search field. The alternatives would have been to have go directly to the site tables and query the columns there (but then we end up with direct binding across services), or to have added stuff to the SiteService API to load distinct property values for a property key.

Growing the SiteService API would have been the cleanest solution, however I’m not sure the extra complexity is worth it for this feature.