SAK-33647 Allow searching for sample users. (#5038)
This allows the SampleUserDirectoryProvider to return users when a search is performed for users. Previously there was no way to search for these users apart from knowing one of the IDs on them.

This makes testing Sakai easier as you can see how it behaves when deployed with a user provider that supports searching.