KNL-1573 Log actual values in the Message (#5069)
* KNL-1573 Log actual values in the Message

When sending a message out we log useful values to aid in debugging. However with the recent re-writing of From: addresses we were logging the original values, rather than the re-written values so the log entries don’t act as a good audit of what got sent.

As we now rely on the replyTo header more because of our re-writing I’ve added that as a header that’s logged as well. This will also help in debugging as it often includes details of the user who performed the action when the from address gets re-written.

Rather than passing in additional values just to be logged we just look directly at the message, this should make it less likely in future that the logged data is different from data used to send the mail.

* Correct formatting and reduce duplication.

(cherry picked from commit 2d98c2908d37b406ee3a910a3797e5053ed11cb7)