SAK-33684 Handle SMTP FROM that isn’t email better (#5080)
Have better handling of a client attempting to send an email where the FROM address isn’t an email address that we can parse. This may be because it’s empty or it’s not an email address.
(cherry picked from commit c05438c0bdf67c51b1890767c35c55f82dc756e5)