SAK-23694 Fail gracefully when role isn’t in group.
When adding a member to a group that belongs to a site, if the role doesn’t exist in the group copy it out of the site and re-attempt the add of the member.

This follows the behaviour of creating the group (KNL-313) where the roles from the site are copied into the group even if they aren’t defined on the group template.

When failing to add a member to a AuthzGroup we give a better error message about what was being attempted.

This also handles the case where the user who is being added isn’t a member of the site any more.