SAK-33466 Fix the course grade mapping
Move the sorting of the grading schema map into the service.
Ensure that the grading schema map is always sorted before applying the course grade lookup mapping so that the mapping applies correctly. This should resolve a long standing bug where course grades may not be mapped correctly or consistently.
Fix a bunch of compiler warnings.