SAKAI-3888 Bug with sections and columns added to blank sections or columns
When adding a section or column after a blank section or column the code searches
for div.item to know which item the new break should be added after. However, there
is no div.item in a blank section or column. So, instead we need to look for a break
item to add the new break after.