Overrode the drawing method for Chartjs so that we can have rounded corners
As documented, this code is based on code by Jed Trow. It has been modified so that only the top (or right, in the case of a horizontal bar) corners are rounded. I have also removed some code that was seemingly redundant.
The github project is linked to in the doco. His code to deal with borderWidth has been left in, in case we want a border on the bars.
The corner radius is defined as half the width or height, whichever is smaller.
Hopefully this feature will be added to a future version of Chartjs and we can remove this method.

Issue: SAK-34043