SAK-31988 removed scrolling when expanding a syllabus item
This change was already made in 0249d876 but got (probably accidentally)
reintroduced in 7a8ef643

It removes the scrolling when clicking on the header of a syllabus item
to expand it. This resulted in a determined scrolling to the bottom of
the page when the "expand all" button was used.